Add files via upload
These files are the most recent and where I am making the changes to the break and fallthru.
Lots of print logging to keep track of everything. Will take out the prints but for now keeping them in to
read the var values.  Still need to change function names and add more documentation.