Add files via upload
Comment out one switch and try one and then try the other.
Numbers and words work now.  I will take out the comment print statements tomorrow
now that it works. What a bear to debug this. So pleased that it actually works and it's fast.
It will be faster after I take out the print statements.