Add files via upload
Still a work in progress but working. Uses reload each time a switch is called (for now).
I am resetting the vars and lists after each switch but for some reason it crashes on switches after first UNLESS I reload the switch module.
With the reload it works like a charm. Will take out the print statements now.
Blake Southwood