Add files via upload
These are two files used as the prototype before I put the methods and functions into
a new module. I used these two files for testing the macro expansion. Sometimes a bug
occurs with more than 15 case sections (working on it) It's a base 10 error regarding an int.
But the code essentially works.