Add files via upload
Working example of three level deep nested switch cases with inner switches.
What is shown here is what the output will be after the python code is generated from
the input strings. I'm still designing the  input flow of control and fuzzy logic.
This proves the logic is sound and the code will work as designed.